+++ df_engine_extensions/src/dist_sql_query/physical_plan.rs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//! dist sql query physical plans
+
 use std::{any::Any, fmt, sync::Arc};
 
 use arrow::datatypes::SchemaRef;
@@ -25,6 +27,7 @@ use datafusion::{
     },
 };
 use table_engine::{
+    provider::ScanTable,
     remote::{model::TableIdentifier, RemoteEngineRef},
     table::ReadRequest,
 };
@@ -164,3 +167,74 @@ impl DisplayAs for ResolvedPartitionedScan {
         )
     }
 }
+
+/// Placeholder of sub table's scan plan
+/// It is inexecutable actually and just for carrying the necessary information
+/// of building the executable scan plan.
+#[derive(Debug)]
+pub struct UnresolvedSubTableScan {
+    pub table: TableIdentifier,
+    pub read_request: ReadRequest,
+}
+
+impl ExecutionPlan for UnresolvedSubTableScan {
+    fn as_any(&self) -> &dyn Any {
+        self
+    }
+
+    fn schema(&self) -> SchemaRef {
+        self.read_request
+            .projected_schema
+            .to_projected_arrow_schema()
+    }
+
+    fn output_partitioning(&self) -> Partitioning {
+        Partitioning::UnknownPartitioning(self.read_request.opts.read_parallelism)
+    }
+
+    fn output_ordering(&self) -> Option<&[PhysicalSortExpr]> {
+        None
+    }
+
+    fn children(&self) -> Vec<Arc<dyn ExecutionPlan>> {
+        vec![]
+    }
+
+    fn with_new_children(
+        self: Arc<Self>,
+        _children: Vec<Arc<dyn ExecutionPlan>>,
+    ) -> DfResult<Arc<dyn ExecutionPlan>> {
+        Err(DataFusionError::Internal(
+            "UnresolvedSubTableScan should not have children".to_string(),
+        ))
+    }
+
+    fn execute(
+        &self,
+        _partition: usize,
+        _context: Arc<TaskContext>,
+    ) -> DfResult<DfSendableRecordBatchStream> {
+        Err(DataFusionError::Internal(
+            "UnresolvedSubTableScan can not be executed".to_string(),
+        ))
+    }
+
+    fn statistics(&self) -> Statistics {
+        Statistics::default()
+    }
+}
+
+impl DisplayAs for UnresolvedSubTableScan {
+    fn fmt_as(&self, _t: DisplayFormatType, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(
+            f,
+            "UnresolvedSubTableScan: table={:?}, read_request:{:?}, partition_count={}",
+            self.table,
+            self.read_request,
+            self.output_partitioning().partition_count(),
+        )
+    }
+}
+
+/// `ResolvedSubTableScan` is `ScanTable` actually.
+pub type ResolvedSubTableScan = ScanTable;
